fixed AuthRequired

+++ src/hoc/AuthRequired.tsx
@@ -13,15 +13,13 @@ const AuthRequired: FC<IProps> = ({children}) => {
     let access = authService.getToken();
     const {currentUser} = useAppSelector(state => state.authReducer);
     const dispatch = useAppDispatch();
-    console.log(access);
-    console.log(currentUser);
     if ((access) && (!currentUser)) {
-    //     if (access) {
-    //         dispatch(authActions.currentUser)
-    //     } else {
+        if (access) {
+            dispatch(authActions.currentUser)
+        } else {
             authService.deleteToken()
             dispatch(authActions.logout)
-        // }
+        }
     }
     const {pathname} = useLocation();
